Add MergePatientsConcern
This adds a concern for jobs which involve merging two patient records
together, either because a patient has been invalidated and given a new
NHS number, or because we already had a patient record for a particular
NHS number.
